+++ test/lib.sh
@@ -442,7 +442,7 @@ function stop_backend(){
 # Wait for restconf to stop sending  502 Bad Gateway
 function wait_backend(){
     freq=$(chunked_framing "<rpc $DEFAULTNS><ping $LIBNS/></rpc>")
-    reply=$(echo "$freq" | $clixon_netconf -qHef $cfg) 
+    reply=$(echo "$freq" | $clixon_netconf -q1ef $cfg) 
 #    freply=$(chunked_framing "<rpc-reply $DEFAULTNS><ok/></rpc-reply>")
 #    chunked_equal "$reply" "$freply"
     let i=0;