+++ CHANGELOG.md
@@ -68,6 +68,9 @@ Users may have to change how they access the system
     * Added 404 return without body if neither restconf, data or streams prefix match
 * Netconf: Usage of chunked framing"
     * To keep existing end-of-message encoding, set `CLICON_NETCONF_BASE_CAPABILITY` to `0`
+    * Added `clixon_netconf` command-line option `-0`and changed `-H` to `-1`
+       * `-0` means dont send hello, but fix netconf base version to 0 and use EOM framing
+       * `-1` means dont send hello, but fix netconf base version to 1 and use chunked framing 
 * New `clixon-config@2022-03-21.yang` revision
   * Added option:
     * `CLICON_RESTCONF_API_ROOT`